+++ src/components/List/List.test.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { listItems } from 'test/fixtures/content.json';
 import List from './List';
+
+const listItems = Array(5).fill().map((val, idx) => `List Item ${idx}`);
 
 describe('List', () => {
   test('should match snapshot', () => {