+++ src/pages/About/About.test.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import { mount } from 'enzyme';
-import AboutPage from './index';
-
-const wrapper = mount(<AboutPage />);
-
-test('AboutPage Page Renders', () => {
-  expect(wrapper).toMatchSnapshot();
-});