+++ src/pages/index.js
@@ -4,3 +4,5 @@ export { default as ButtonsPage } from './Buttons';
 export { default as CardsPage } from './Cards';
 export { default as TodoPage } from './Todos';
 export { default as ComponentsPage } from './Components';
+export { default as BigListPage } from './BigList';
+export { default as BigTablePage } from './BigTable';